Added new Bass Squeezer plugin
Split-band compression and filtered distortion for that instant bathtub sound. Mostly intended for electric basses in modern rock, amped or DI. Compress the low end until it blows your sub, distort the high end until it melts your face, and nobody wants the low mids in a bass anyway, so this plugin can conveniently lose them for you as well. No more excuses for a weak or unstable low end in your mix.