Added new Chug Thug plugin
Simplified split-band processor to tame the low end of distorted guitars when palm muting and chugging. The Crossover frequency defines how much low end will be split off and processed. The Amount slider controls how hard the signal is driven into the compressor, this interacts with the Input gain. Intended for use on mono, dual mono or stereo recordings of already amplified guitars.